test(booking-service): add WebMvc tests for booking details endpoint and format imports

+++ booking-service/src/main/java/com/team28/booking/booking/dto/BookingDetailsDTO.java
@@ -7,14 +7,13 @@ import java.util.List;
 import java.util.Map;
 
 public record BookingDetailsDTO(
-        Long bookingId,
-        Long userId,
-        Long providerId,
-        Booking.Status status,
-        BigDecimal totalPrice,
-        Map<String, Object> metadata,
-        List<ServiceDetailsDTO> services,
-        Integer totalServices,
-        Long completedServices
-) {
+                Long bookingId,
+                Long userId,
+                Long providerId,
+                Booking.Status status,
+                BigDecimal totalPrice,
+                Map<String, Object> metadata,
+                List<ServiceDetailsDTO> services,
+                Integer totalServices,
+                Integer completedServices) {
 }
